requirements.txt
ansible-core>=2.17.2,<3.0.0
omsdk
redfish
molecule>=3.0.0,<4.0.0
molecule-docker>=1.0.0,<3.0.0
docker>=4.3.1,<5.0.0